requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.13.2
amqp==5.1.1
anyio==3.6.2
apache-airflow==2.9.3
apache-airflow-providers-apache-spark==4.9.0
apache-airflow-providers-celery==3.1.0
apache-airflow-providers-common-io==1.4.0
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-fab==1.2.2
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-postgres==5.4.0
apache-airflow-providers-redis==3.1.0
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.3.2
apispec==6.6.1
argcomplete==3.0.8
asgiref==3.6.0
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
billiard==3.6.4.0
blinker==1.6.2
cachelib==0.9.0
cattrs==22.2.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
ConfigUpdater==3.1.1
connexion==2.14.2
cramjam==2.8.3
cron_descriptor==1.2.35
croniter==3.0.3
cryptography==40.0.2
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.3.0
docutils==0.19
duckdb==1.0.0
duckdb_engine==0.13.1
email-validator==1.3.1
exceptiongroup==1.1.1
executing==2.0.1
fastparquet==2024.5.0
Flask==2.2.4
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.0
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
flower==1.2.0
frozenlist==1.3.3
fsspec==2024.6.1
google-re2==1.1.20240702
googleapis-common-protos==1.63.2
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.65.4
grpcio-status==1.62.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
humanize==4.6.0
idna==3.4
importlib-resources==5.12.0
importlib_metadata==8.0.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kombu==5.2.4
lazy-object-proxy==1.9.0
limits==3.4.0
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.5
mdurl==0.1.2
methodtools==0.4.7
multidict==6.0.4
nest-asyncio==1.6.0
numpy==2.0.1
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
ordered-set==4.1.0
packaging==21.3
pandas==2.2.2
parso==0.8.4
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.9.0
platformdirs==4.2.2
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.25.4
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.5
pyarrow==17.0.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
pyspark==3.3.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==26.1.0
redis==3.5.3
referencing==0.35.1
requests==2.29.0
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.3.5
rich_argparse==1.1.0
rpds-py==0.20.0
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
tornado==6.3.1
traitlets==5.14.3
typing_extensions==4.5.0
tzdata==2024.1
uc-micro-py==1.0.1
unicodecsv==0.14.1
Unidecode==1.3.8
universal_pathlib==0.2.2
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.19.2
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.13.2
amqp==5.1.1
anyio==3.6.2
apache-airflow==2.9.3
apache-airflow-providers-apache-spark==4.9.0
apache-airflow-providers-celery==3.1.0
apache-airflow-providers-common-io==1.4.0
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-fab==1.2.2
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-postgres==5.4.0
apache-airflow-providers-redis==3.1.0
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.3.2
apispec==6.6.1
argcomplete==3.0.8
asgiref==3.6.0
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
billiard==3.6.4.0
blinker==1.6.2
cachelib==0.9.0
cattrs==22.2.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
ConfigUpdater==3.1.1
connexion==2.14.2
cramjam==2.8.3
cron_descriptor==1.2.35
croniter==3.0.3
cryptography==40.0.2
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.3.0
docutils==0.19
duckdb==1.0.0
duckdb_engine==0.13.1
email-validator==1.3.1
exceptiongroup==1.1.1
executing==2.0.1
fastparquet==2024.5.0
Flask==2.2.4
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.0
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
flower==1.2.0
frozenlist==1.3.3
fsspec==2024.6.1
google-re2==1.1.20240702
googleapis-common-protos==1.63.2
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.65.4
grpcio-status==1.62.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
humanize==4.6.0
idna==3.4
importlib-resources==5.12.0
importlib_metadata==8.0.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kombu==5.2.4
lazy-object-proxy==1.9.0
limits==3.4.0
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.5
mdurl==0.1.2
methodtools==0.4.7
multidict==6.0.4
nest-asyncio==1.6.0
numpy==2.0.1
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
ordered-set==4.1.0
packaging==21.3
pandas==2.2.2
parso==0.8.4
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.9.0
platformdirs==4.2.2
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.25.4
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.5
pyarrow==17.0.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
pyspark==3.3.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==26.1.0
redis==3.5.3
referencing==0.35.1
requests==2.29.0
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.3.5
rich_argparse==1.1.0
rpds-py==0.20.0
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
tornado==6.3.1
traitlets==5.14.3
typing_extensions==4.5.0
tzdata==2024.1
uc-micro-py==1.0.1
unicodecsv==0.14.1
Unidecode==1.3.8
universal_pathlib==0.2.2
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.19.2
